Compilation fixes.
