--- core/designsystem/src/commonMain/kotlin/com/adammcneilly/spacenerd/core/designsystem/SpaceNerdColors.kt
@@ -1,10 +1,0 @@
-@file:Suppress("MagicNumber")
-
-package com.adammcneilly.spacenerd.core.designsystem
-
-import androidx.compose.ui.graphics.Color
-
-object SpaceNerdColors {
-    val Green = Color(0xFF4CAF50)
-    val SeedColor = Color(0xFF121212)
-}
