+++ core/designsystem/src/commonMain/kotlin/com/adammcneilly/spacenerd/core/designsystem/theme/SpaceNerdColors.kt
@@ -1,0 +1,13 @@
+@file:Suppress("MagicNumber")
+
+package com.adammcneilly.spacenerd.core.designsystem.theme
+
+import androidx.compose.ui.graphics.Color
+
+/**
+ * Entire color palette generated by Gemini. Take it with a grain of salt.
+ */
+object SpaceNerdColors {
+    val Green = Color(0xFF4CAF50)
+    val SeedColor = Color(0xFF121212)
+}
