+++ core/designsystem/src/commonMain/kotlin/com/adammcneilly/spacenerd/core/designsystem/utils/AnimationScopes.kt
@@ -1,0 +1,15 @@
+package com.adammcneilly.spacenerd.core.designsystem.utils
+
+import androidx.compose.animation.AnimatedVisibilityScope
+import androidx.compose.animation.ExperimentalSharedTransitionApi
+import androidx.compose.animation.SharedTransitionScope
+import androidx.compose.runtime.compositionLocalOf
+
+@OptIn(ExperimentalSharedTransitionApi::class)
+val LocalSharedTransitionScope = compositionLocalOf<SharedTransitionScope?> {
+    null
+}
+
+val LocalNavAnimatedVisibilityScope = compositionLocalOf<AnimatedVisibilityScope?> {
+    null
+}
