+++ core/designsystem/src/commonMain/kotlin/com/adammcneilly/spacenerd/core/designsystem/utils/SceneType.kt
@@ -1,0 +1,12 @@
+package com.adammcneilly.spacenerd.core.designsystem.utils
+
+import androidx.compose.runtime.compositionLocalOf
+
+val LocalSceneType = compositionLocalOf {
+    SceneType.Single
+}
+
+enum class SceneType {
+    Single,
+    TwoPane,
+}
