+++ core/displaymodels/src/commonMain/kotlin/com/adammcneilly/spacenerd/core/displaymodels/LaunchStatusDisplayModel.kt
@@ -1,7 +1,7 @@
 package com.adammcneilly.spacenerd.core.displaymodels
 
 import androidx.compose.ui.graphics.Color
-import com.adammcneilly.spacenerd.core.designsystem.SpaceNerdColors
+import com.adammcneilly.spacenerd.core.designsystem.theme.SpaceNerdColors
 import com.adammcneilly.spacenerd.core.models.LaunchStatus
 
 data class LaunchStatusDisplayModel(
