+++ shared/build.gradle.kts
@@ -46,6 +46,7 @@ kotlin {
 
         commonMain.dependencies {
             implementation(project(":core:datetime"))
+            implementation(project(":core:designsystem"))
             implementation(project(":core:displaymodels"))
             implementation(project(":core:models"))
             implementation(project(":data:agency"))
