+++ shared/src/commonMain/kotlin/com/adammcneilly/spacenerd/shared/app/App.kt
@@ -7,9 +7,9 @@ import androidx.compose.material3.Surface
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.ui.Modifier
+import com.adammcneilly.spacenerd.core.designsystem.theme.SpaceTheme
 import com.adammcneilly.spacenerd.shared.di.appModules
 import com.adammcneilly.spacenerd.shared.scaffold.LocalSharedTransitionScope
-import com.adammcneilly.spacenerd.shared.ui.theme.SpaceTheme
 import org.koin.compose.KoinApplication
 
 @Composable
