+++ shared/src/commonMain/kotlin/com/adammcneilly/spacenerd/shared/feature/launchdetail/ui/LaunchDetailContent.kt
@@ -18,11 +18,11 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.unit.dp
+import com.adammcneilly.spacenerd.core.designsystem.utils.sharedBounds
+import com.adammcneilly.spacenerd.core.designsystem.utils.sharedElement
 import com.adammcneilly.spacenerd.core.displaymodels.LaunchDisplayModel
 import com.adammcneilly.spacenerd.shared.ui.components.ImageWrapper
 import com.adammcneilly.spacenerd.shared.ui.components.Pill
-import com.adammcneilly.spacenerd.shared.ui.utils.sharedBounds
-import com.adammcneilly.spacenerd.shared.ui.utils.sharedElement
 import com.eygraber.compose.placeholder.PlaceholderDefaults
 import com.eygraber.compose.placeholder.material3.color
 import com.eygraber.compose.placeholder.material3.placeholder
