--- shared/src/commonMain/kotlin/com/adammcneilly/spacenerd/shared/ui/theme/Shape.kt
@@ -1,11 +1,0 @@
-package com.adammcneilly.spacenerd.shared.ui.theme
-
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.Shapes
-import androidx.compose.ui.unit.dp
-
-val Shapes = Shapes(
-    small = RoundedCornerShape(4.dp),
-    medium = RoundedCornerShape(8.dp),
-    large = RoundedCornerShape(16.dp),
-)
