--- shared/src/commonMain/kotlin/com/adammcneilly/spacenerd/shared/ui/theme/SpaceNerdColors.kt
@@ -1,13 +1,0 @@
-@file:Suppress("MagicNumber")
-
-package com.adammcneilly.spacenerd.shared.ui.theme
-
-import androidx.compose.ui.graphics.Color
-
-/**
- * Entire color palette generated by Gemini. Take it with a grain of salt.
- */
-object SpaceNerdColors {
-    val Green = Color(0xFF4CAF50)
-    val SeedColor = Color(0xFF121212)
-}
