--- shared/src/commonMain/kotlin/com/adammcneilly/spacenerd/shared/ui/utils/SceneType.kt
@@ -1,12 +1,0 @@
-package com.adammcneilly.spacenerd.shared.ui.utils
-
-import androidx.compose.runtime.compositionLocalOf
-
-val LocalSceneType = compositionLocalOf {
-    SceneType.Single
-}
-
-enum class SceneType {
-    Single,
-    TwoPane,
-}
