--- shared/src/iosMain/kotlin/com/adammcneilly/spacenerd/shared/ui/utils/WindowSizeClass.ios.kt
@@ -1,12 +1,0 @@
-package com.adammcneilly.spacenerd.shared.ui.utils
-
-import androidx.compose.material3.windowsizeclass.ExperimentalMaterial3WindowSizeClassApi
-import androidx.compose.material3.windowsizeclass.WindowSizeClass
-import androidx.compose.material3.windowsizeclass.calculateWindowSizeClass
-import androidx.compose.runtime.Composable
-
-@Composable
-@OptIn(ExperimentalMaterial3WindowSizeClassApi::class)
-actual fun currentWindowSizeClass(): WindowSizeClass {
-    return calculateWindowSizeClass()
-}
